requirements-dev.txt
colorama~=0.4.6
jwcrypto~=1.5.6
pyjwt~=2.8.0
requests~=2.31.0

iniconfig~=2.0.0
pluggy~=1.4.0
pytest~=8.1.1
black~=24.3